+++ cups/testraster.c
@@ -1,7 +1,7 @@
 /*
  * Raster test program routines for CUPS.
  *
- * Copyright © 2007-2018 by Apple Inc.
+ * Copyright © 2007-2019 by Apple Inc.
  * Copyright © 1997-2007 by Easy Software Products.
  *
  * Licensed under Apache License v2.0.  See the file "LICENSE" for more
@@ -22,8 +22,7 @@
 
 static int	do_ras_file(const char *filename);
 static int	do_raster_tests(cups_mode_t mode);
-static void	print_changes(cups_page_header2_t *header,
-		              cups_page_header2_t *expected);
+static void	print_changes(cups_page_header2_t *header, cups_page_header2_t *expected);
 
 
 /*
@@ -174,6 +173,8 @@ do_raster_tests(cups_mode_t mode)	/* O - Write mode */
     header.cupsPageSize[0]  = 288.0f;
     header.cupsPageSize[1]  = 288.0f;
 
+    strlcpy(header.MediaType, "auto", sizeof(header.MediaType));
+
     if (page & 1)
     {
       header.cupsBytesPerLine *= 4;
@@ -307,6 +308,8 @@ do_raster_tests(cups_mode_t mode)	/* O - Write mode */
     expected.HWResolution[1]  = 64;
     expected.PageSize[0]      = 288;
     expected.PageSize[1]      = 288;
+
+    strlcpy(expected.MediaType, "auto", sizeof(expected.MediaType));
 
     if (mode != CUPS_RASTER_WRITE_PWG)
     {